[Lookup] Test data should also be read as float32 container type

+++ src/finn/custom_op/fpgadataflow/hls/lookup_hls.py
@@ -98,7 +98,7 @@ class Lookup_hls(Lookup, HLSBackend):
         packed_bits = self.get_instream_width()
         packed_hls_type = "ap_uint<%d>" % packed_bits
         elem_hls_type = dtype.get_hls_datatype_str()
-        npy_type = "int64_t"
+        npy_type = "float"
         npy_in = "%s/input_0.npy" % code_gen_dir
         self.code_gen_dict["$READNPYDATA$"] = []
         self.code_gen_dict["$READNPYDATA$"].append(
